client/python/requirements.txt
etcd3==0.12.0
grpcio==1.32.0